remove note from README.TXT about problems that are gone now
Running without JIT no longer seems necessary: The bug was indeed
zigbee-lua's use of a *threaded* API of the MQTT library without
any synchronization mechanism (which is not directly possible using
just the FFI). Switching over to a different API got rid of the
strange behaviour which doesn't look that strange in hindsight.