added new "httpd10" interface: pure LuaJIT+LJSyscall httpd server
This is to get rid of the instable libmicrohttpd binding, which works
fine most of the time - and then crashes LuaJIT due to threading and
wacky callbacks between Lua and C.

The new httpd server interface is *very* simple. It claims that it
is HTTP 1.0, which is probably not really true, as it still misses
lots of stuff. But it is small and concise and it works happily
behind an nginx frontend server for me - in fact, it does talk to
browsers, too. You should probably not offer any public access to
it. SSL/TLS will never be implemented, use a frontend proxy for
this. There will probably be bugs, security relevant ones at that.
Resource exhaustion is simple with this one.